Correct naming mistakes
